requirements.txt
mlflow>=2.3.0
evidently>=0.2.8
pydantic>=1.10.0
numpy>=1.21.0
PyYAML>=6.0
scikit-learn>=1.0.2